scaffolder-backend: move catalog:register action from legacy to builtin
